move refactor a bunch of api's
