+++ app/api/v1/models/__init__.py
@@ -4,6 +4,7 @@ from .decrypt import *
 from .encrypt import *
 from .election import *
 from .guardian import *
-from .key import *
+from .key_ceremony import *
+from .key_guardian import *
 from .manifest import *
 from .tally import *
