+++ app/api/v1/models/key_guardian.py
@@ -12,37 +12,16 @@ __all__ = [
     "GuardianQueryResponse",
     "GuardianSubmitVerificationRequest",
     "GuardianSubmitChallengeRequest",
-    "KeyCeremony",
-    "KeyCeremonyState",
     "KeyCeremonyGuardian",
     "KeyCeremonyGuardianStatus",
     "KeyCeremonyGuardianState",
-    "KeyCeremonyCreateRequest",
-    "KeyCeremonyStateResponse",
-    "KeyCeremonyQueryResponse",
-    "KeyCeremonyVerifyChallengesResponse",
-    "ElectionJointKeyResponse",
 ]
 
-ElectionPublicKey = Any
-ElGamalKeyPair = Any
 PublicKeySet = Any
 
 ElectionPartialKeyBackup = Any
 ElectionPartialKeyVerification = Any
 ElectionPartialKeyChallenge = Any
-
-ElectionJointKey = Any
-
-
-class KeyCeremonyState(str, Enum):
-    """Enumeration expressing the state of the key caremony."""
-
-    CREATED = "CREATED"
-    OPEN = "OPEN"
-    CLOSED = "CLOSED"
-    CHALLENGED = "CHALLENGED"
-    CANCELLED = "CANCELLED"
 
 
 class KeyCeremonyGuardianStatus(str, Enum):
@@ -76,32 +55,6 @@ class KeyCeremonyGuardian(Base):
     challenges: Optional[List[ElectionPartialKeyChallenge]]
 
 
-class KeyCeremony(Base):
-    """The Key Ceremony is a record of the state of a key ceremony."""
-
-    key_name: str
-    state: KeyCeremonyState
-    number_of_guardians: int
-    quorum: int
-    guardian_ids: List[GUARDIAN_ID]
-    guardian_status: Dict[GUARDIAN_ID, KeyCeremonyGuardianState]
-    election_joint_key: Optional[ElectionJointKey]
-
-
-class KeyCeremonyStateResponse(Base):
-    key_name: str
-    state: KeyCeremonyState
-    guardian_status: Dict[GUARDIAN_ID, KeyCeremonyGuardianState]
-
-
-class KeyCeremonyQueryResponse(BaseResponse):
-    key_ceremonies: List[KeyCeremony]
-
-
-class KeyCeremonyVerifyChallengesResponse(BaseResponse):
-    verifications: List[ElectionPartialKeyVerification]
-
-
 class GuardianAnnounceRequest(BaseRequest):
     """A set of public auxiliary and election keys"""
 
@@ -130,33 +83,3 @@ class GuardianSubmitChallengeRequest(BaseRequest):
     key_name: str
     guardian_id: str
     challenges: List[ElectionPartialKeyChallenge]
-
-
-class AuxiliaryPublicKeyRequest(BaseRequest):
-    """Auxiliary public key and owner information"""
-
-    owner_id: str
-    sequence_order: int
-    key: str
-
-
-class KeyCeremonyCreateRequest(BaseRequest):
-    """Request to create a new key ceremony"""
-
-    key_name: str
-    number_of_guardians: int
-    quorum: int
-    guardian_ids: List[str]
-
-
-class PublishElectionJointKeyRequest(BaseRequest):
-    """Request to publish the election joint key"""
-
-    key_name: str
-    election_public_keys: List[ElectionPublicKey]
-
-
-class ElectionJointKeyResponse(BaseResponse):
-    """Response object containing the Election Joint Key"""
-
-    joint_key: ElectionJointKey
