+++ app/api/v1/tags.py
@@ -1,6 +1,8 @@
 ELECTION = "Configure Election"
 MANIFEST = "Election Manifest"
+GUARDIAN = "Guardian"
 KEY_CEREMONY = "Key Ceremony"
+KEY_GUARDIAN = "Key Ceremony Participant"
 BALLOTS = "Query Ballots"
 ENCRYPT = "Encrypt Ballots"
 TALLY = "Tally Results"
