+++ app/core/client.py
@@ -1,0 +1,6 @@
+# TODO: multi-tenancy
+DEFAULT_CLIENT_ID = "electionguard-default-client-id"
+
+
+def get_client_id() -> str:
+    return DEFAULT_CLIENT_ID
